requirements.txt
fastapi==0.109.0
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.9.0.80
uvicorn==0.27.0.post1
python-dotenv
openai==0.28
boto3==1.28.57
